+++ pkg/ingester/active_series.go
@@ -21,8 +21,9 @@ const activeSeriesMaxSizeBytes = 1 * 1024 * 1024
 
 // ActiveSeries implements the ActiveSeries RPC. It returns a stream of active
 // series that match the given matchers.
-func (i *Ingester) ActiveSeries(request *client.ActiveSeriesRequest, stream client.Ingester_ActiveSeriesServer) error {
-	if err := i.checkRunning(); err != nil {
+func (i *Ingester) ActiveSeries(request *client.ActiveSeriesRequest, stream client.Ingester_ActiveSeriesServer) (asErr error) {
+	defer func() { asErr = i.handleReadError(asErr) }()
+	if err := i.checkAvailable(); err != nil {
 		return err
 	}
 	if err := i.checkReadOverloaded(); err != nil {